+++ src/components/Auth/AuthInput/AuthInputBox/AuthInputItem/ErrorMessage.tsx
@@ -1,9 +1,11 @@
+import { ReactNode } from 'react';
+
 interface Props {
-  children: string;
+  children: ReactNode;
 }
 
 const ErrorMessage = ({ children }: Props) => {
-  return <p className="body6 text-red">{children}</p>;
+  return <p className="body6 mt-1 text-red">{children}</p>;
 };
 
 export default ErrorMessage;
