+++ src/components/Auth/AuthInput/AuthInputBox/index.ts
@@ -1,5 +1,13 @@
 import AuthEmailInputBox from './AuthEmailInputBox';
 import AuthPwInputBox from './AuthPwInputBox';
 import AuthPwCheckInputBox from './AuthPwCheckInputBox';
+import { AuthInput, AuthInputWrapper, ErrorMessage } from './AuthInputItem';
 
-export { AuthEmailInputBox, AuthPwInputBox, AuthPwCheckInputBox };
+export {
+  AuthEmailInputBox,
+  AuthPwInputBox,
+  AuthPwCheckInputBox,
+  AuthInput,
+  AuthInputWrapper,
+  ErrorMessage,
+};
