+++ src/components/Auth/Login/index.ts
@@ -1,0 +1,5 @@
+import LoginLogo from './LoginLogo';
+import LoginForm from './LoginForm';
+import AuthButtonsWrapper from './AuthButtons/AuthButtonsWrapper';
+
+export { LoginLogo, LoginForm, AuthButtonsWrapper };
